mysite/requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astor==0.8.1
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Distance==0.1.3
Django==3.1.4
filelock==3.0.12
gast==0.2.2
gdown==3.12.2
google-pasta==0.2.0
googledrivedownloader==0.4
grpcio==1.34.0
h5py==3.1.0
idna==2.10
importlib-metadata==3.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.2
Markdown==3.3.3
matplotlib==3.2.0
numpy==1.19.4
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.4.0.46
opt-einsum==3.3.0
Pillow @ file:///D:/bld/pillow_1604749014518/work
protobuf==3.14.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
tesseract==0.1.3
tesserocr @ file:///D:/bld/tesserocr_1602669179082/work
tqdm==4.54.1
urllib3==1.26.2
Wand==0.6.5
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.0
